requirements.txt
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.12
greenlet==3.1.1
groq==0.20.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.21
langchain-core==0.3.49
langchain-text-splitters==0.3.7
langchainhub==0.1.21
langsmith==0.3.19
orjson==3.10.16
packaging==24.2
pika==1.3.2
psutil==7.0.0
pydantic==2.11.1
pydantic_core==2.33.0
python-dotenv==1.1.0
pywin32==310
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
tenacity==9.0.0
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
uvicorn==0.34.0
zstandard==0.23.0
